(feat.) fix post comment button
the reason of why post button is not working is because it is from `editor.isFocused`.

when click event is happening outside of the dom of the `editor` the browser will just lose focusing on the editor but not focus on another dom (button have click the browser ignore it).